SDIT-2620 Remove old index code - refresh event fix
