+++ hmpps-prisoner-search-indexer/src/test/kotlin/uk/gov/justice/digital/hmpps/prisonersearch/indexer/listeners/DomainEventPublisherListenerTest.kt
@@ -1,7 +1,5 @@
 package uk.gov.justice.digital.hmpps.prisonersearch.indexer.listeners
 
-import com.fasterxml.jackson.databind.ObjectMapper
-import kotlinx.coroutines.test.runTest
 import org.awaitility.kotlin.await
 import org.awaitility.kotlin.matches
 import org.awaitility.kotlin.untilCallTo
@@ -11,7 +9,6 @@ import org.mockito.kotlin.any
 import org.mockito.kotlin.doReturn
 import org.mockito.kotlin.doThrow
 import org.mockito.kotlin.whenever
-import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.beans.factory.annotation.Qualifier
 import org.springframework.test.context.bean.override.mockito.MockitoSpyBean
 import software.amazon.awssdk.services.sns.SnsAsyncClient
@@ -20,9 +17,6 @@ import software.amazon.awssdk.services.sns.model.PublishResponse
 import software.amazon.awssdk.services.sqs.model.SendMessageRequest
 import uk.gov.justice.digital.hmpps.prisonersearch.indexer.IntegrationTestBase
 import uk.gov.justice.hmpps.sqs.HmppsQueue
-import uk.gov.justice.hmpps.sqs.MissingQueueException
-import uk.gov.justice.hmpps.sqs.PurgeQueueRequest
-import uk.gov.justice.hmpps.sqs.countAllMessagesOnQueue
 import java.util.concurrent.CompletableFuture
 
 class DomainEventPublisherListenerTest : IntegrationTestBase() {
@@ -31,17 +25,9 @@ class DomainEventPublisherListenerTest : IntegrationTestBase() {
   @Qualifier("hmppseventtopic-sns-client")
   private lateinit var hmppsEventTopicSnsClient: SnsAsyncClient
 
-  @Autowired
-  private lateinit var objectMapper: ObjectMapper
-
   private val publishQueue by lazy { hmppsQueueService.findByQueueId("publish") as HmppsQueue }
   private val publishSqsClient by lazy { publishQueue.sqsClient }
   private val publishQueueUrl by lazy { publishQueue.queueUrl }
-
-  private val hmppsEventsQueue by lazy {
-    hmppsQueueService.findByQueueId("hmppseventtestqueue")
-      ?: throw MissingQueueException("hmppseventtestqueue queue not found")
-  }
 
   private val domainEventPublishRequest by lazy {
     SendMessageRequest.builder().queueUrl(publishQueueUrl)
@@ -57,13 +43,7 @@ class DomainEventPublisherListenerTest : IntegrationTestBase() {
   }
 
   @BeforeEach
-  fun purgeHmppsEventsQueue() = runTest {
-    with(hmppsEventsQueue) {
-      hmppsQueueService.purgeQueue(PurgeQueueRequest(queueName, sqsClient, queueUrl))
-    }
-  }
-
-  private fun getNumberOfMessagesCurrentlyOnDomainQueue(): Int? = hmppsEventsQueue.sqsClient.countAllMessagesOnQueue(hmppsEventsQueue.queueUrl).get()
+  fun purgeHmppsEventsQueue() = purgeDomainEventsQueue()
 
   @Test
   fun `can publish a message`() {
