+++ hmpps-prisoner-search-indexer/src/test/kotlin/uk/gov/justice/digital/hmpps/prisonersearch/indexer/resource/DomainEventsResourceIntTest.kt
@@ -2,42 +2,20 @@
 
 package uk.gov.justice.digital.hmpps.prisonersearch.indexer.resource
 
-import com.fasterxml.jackson.databind.ObjectMapper
-import com.fasterxml.jackson.module.kotlin.readValue
 import kotlinx.coroutines.ExperimentalCoroutinesApi
-import kotlinx.coroutines.test.runTest
 import net.javacrumbs.jsonunit.assertj.JsonAssertions.assertThatJson
 import org.awaitility.kotlin.await
 import org.awaitility.kotlin.matches
 import org.awaitility.kotlin.untilCallTo
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
-import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.http.MediaType
 import org.springframework.web.reactive.function.BodyInserters
-import software.amazon.awssdk.services.sqs.model.DeleteMessageRequest
-import software.amazon.awssdk.services.sqs.model.ReceiveMessageRequest
 import uk.gov.justice.digital.hmpps.prisonersearch.indexer.IntegrationTestBase
-import uk.gov.justice.digital.hmpps.prisonersearch.indexer.services.events.MsgBody
-import uk.gov.justice.hmpps.sqs.MissingQueueException
-import uk.gov.justice.hmpps.sqs.PurgeQueueRequest
-import uk.gov.justice.hmpps.sqs.countAllMessagesOnQueue
 
 class DomainEventsResourceIntTest : IntegrationTestBase() {
-  @Autowired
-  private lateinit var objectMapper: ObjectMapper
-
-  private val hmppsEventsQueue by lazy {
-    hmppsQueueService.findByQueueId("hmppseventtestqueue")
-      ?: throw MissingQueueException("hmppseventtestqueue queue not found")
-  }
-
   @BeforeEach
-  fun purgeHmppsEventsQueue() = runTest {
-    with(hmppsEventsQueue) {
-      hmppsQueueService.purgeQueue(PurgeQueueRequest(queueName, sqsClient, queueUrl))
-    }
-  }
+  fun purgeHmppsEventsQueue() = purgeDomainEventsQueue()
 
   @Test
   fun `access forbidden when no authority`() {
@@ -183,16 +161,4 @@ class DomainEventsResourceIntTest : IntegrationTestBase() {
     assertThatJson(message).node("additionalInformation.nomsNumber").isEqualTo("A2483AA")
     assertThatJson(message).node("additionalInformation.reason").isEqualTo("TRANSFERRED")
   }
-
-  private fun readNextDomainEventMessage(): String {
-    val updateResult = hmppsEventsQueue.sqsClient.receiveMessage(
-      ReceiveMessageRequest.builder().queueUrl(hmppsEventsQueue.queueUrl).build(),
-    ).get().messages().first()
-    hmppsEventsQueue.sqsClient.deleteMessage(
-      DeleteMessageRequest.builder().queueUrl(hmppsEventsQueue.queueUrl).receiptHandle(updateResult.receiptHandle()).build(),
-    ).get()
-    return objectMapper.readValue<MsgBody>(updateResult.body()).Message
-  }
-
-  private fun getNumberOfMessagesCurrentlyOnDomainQueue(): Int? = hmppsEventsQueue.sqsClient.countAllMessagesOnQueue(hmppsEventsQueue.queueUrl).get()
 }
